Removed SendToNVM checkbox and added a store and select configuration button for the NVM. Added file for handling text inputs in message box. Added command as an argument for SendSC.
